requirements.txt

---

### **requirements.txt**

```plaintext
aiohttp==3.8.1
requests==2.28.1
termcolor==2.2.0
dnspython==2.2.1
